requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.3
django-extensions==2.2.8
djangorestframework==3.11.0
elasticsearch==7.5.1
idna==2.6
psycopg2==2.8.4
pycparser==2.19
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.18.4
six==1.14.0
sqlparse==0.3.0
urllib3==1.22
